Renamed reader_serial_port_number() into reader_serial_get_reader_type(). Changed the way CT-API handles the reader_type : removed the porking through the port number value. Should work, to be tested ;)
git-svn-id: https://svn.streamboard.tv/oscam/branches/merlin@257 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ csctapi/ctapi.c
@@ -50,7 +50,7 @@ static pthread_mutex_t ct_list_mutex = PTHREAD_MUTEX_INITIALIZER;
  * Exported functions definition
  */
 
-char CT_init(unsigned short ctn, unsigned short pn, int reader_type)
+char CT_init(unsigned short ctn, unsigned short pn, unsigned short reader_type)
 {
 	CardTerminal *ct;
 	char ret;
